test(metadata): move Dynamo value generation convention coverage
